Extract tests for isInt16Array to own file
